Misc Java 11 cleanup (#958)

+++ src/main/java/org/jenkinsci/test/acceptance/utils/aether/ConsoleTransferListener.java
@@ -57,14 +57,14 @@ public class ConsoleTransferListener
     public void transferProgressed( TransferEvent event )
     {
         TransferResource resource = event.getResource();
-        downloads.put( resource, Long.valueOf( event.getTransferredBytes() ) );
+        downloads.put( resource, event.getTransferredBytes());
 
         StringBuilder buffer = new StringBuilder( 64 );
 
         for ( Map.Entry<TransferResource, Long> entry : downloads.entrySet() )
         {
             long total = entry.getKey().getContentLength();
-            long complete = entry.getValue().longValue();
+            long complete = entry.getValue();
 
             buffer.append( getStatus( complete, total ) ).append( "  " );
         }
